[Feat] Extract RGBColorInProgressCircle to file
