+++ ToGather/ToGather/Screens/ProgressCircle/ProgressCircle.swift
@@ -7,34 +7,6 @@
 
 import SwiftUI
 
-// 그라데이션 컬러
-struct RGBColorInProgressCircle {
-    let red: Double
-    let green: Double
-    let blue: Double
-
-    let multipleRed = 0.61
-    let multipleGreen = 0.62
-    let multipleBlue = 0.93
-
-    var start: Color {
-        return Color(red: red, green: green, blue: blue)
-    }
-
-    var end: Color {
-        return Color(red: red * multipleRed, green: green * multipleGreen, blue: blue * multipleBlue)
-    }
-
-    var background: Color {
-        return start.opacity(0.05)
-    }
-    var line: Color {
-        return start.opacity(0.20)
-    }
-}
-
-let myColor = RGBColorInProgressCircle(red: 0.44, green: 0.66, blue: 0.97)
-
 var savePeriod: Int {
     return saveList.count
 }
